# 🧹 chore: Improve Performance of Fiber Router