--- a/app.go
+++ b/app.go
@@ -872,11 +872,11 @@ func (app *App) Config() Config {
 func (app *App) Handler() fasthttp.RequestHandler { //revive:disable-line:confusing-naming // Having both a Handler() (uppercase) and a handler() (lowercase) is fine. TODO: Use nolint:revive directive instead. See https://github.com/golangci/golangci-lint/issues/3476
 	// prepare the server for the start
 	app.startupProcess()
+
 	if app.newCtxFunc != nil {
 		return app.customRequestHandler
-	} else {
-		return app.defaultRequestHandler
 	}
+	return app.defaultRequestHandler
 }
 
 // Stack returns the raw router stack.