--- a/app_test.go
+++ b/app_test.go
@@ -581,11 +581,6 @@ func Test_App_Use_StrictRouting(t *testing.T) {
 
 func Test_App_Add_Method_Test(t *testing.T) {
 	t.Parallel()
-	defer func() {
-		if err := recover(); err != nil {
-			require.Equal(t, "add: invalid http method JANE\n", fmt.Sprintf("%v", err))
-		}
-	}()
 
 	methods := append(DefaultMethods, "JOHN") //nolint:gocritic // We want a new slice here
 	app := New(Config{
@@ -607,11 +602,9 @@ func Test_App_Add_Method_Test(t *testing.T) {
 	require.Equal(t, StatusNotImplemented, resp.StatusCode, "Status code")
 
 	// Add a new method
-	app.Add([]string{"JANE"}, "/jane", testEmptyHandler)
-
-	resp, err = app.Test(httptest.NewRequest("JANE", "/jane", nil))
-	require.NoError(t, err, "app.Test(req)")
-	require.Equal(t, StatusOK, resp.StatusCode, "Status code")
+	require.Panics(t, func() {
+		app.Add([]string{"JANE"}, "/jane", testEmptyHandler)
+	})
 }
 
 func Test_App_All_Method_Test(t *testing.T) {