--- a/app_test.go
+++ b/app_test.go
@@ -609,11 +609,6 @@ func Test_App_Add_Method_Test(t *testing.T) {
 
 func Test_App_All_Method_Test(t *testing.T) {
 	t.Parallel()
-	defer func() {
-		if err := recover(); err != nil {
-			require.Equal(t, "add: invalid http method JANE\n", fmt.Sprintf("%v", err))
-		}
-	}()
 
 	methods := append(DefaultMethods, "JOHN") //nolint:gocritic // We want a new slice here
 	app := New(Config{
@@ -628,11 +623,9 @@ func Test_App_All_Method_Test(t *testing.T) {
 	require.Equal(t, StatusOK, resp.StatusCode, "Status code")
 
 	// Add a new method
-	app.Add([]string{"JANE"}, "/doe", testEmptyHandler)
-
-	resp, err = app.Test(httptest.NewRequest("JANE", "/doe", nil))
-	require.NoError(t, err, "app.Test(req)")
-	require.Equal(t, StatusOK, resp.StatusCode, "Status code")
+	require.Panics(t, func() {
+		app.Add([]string{"JANE"}, "/jane", testEmptyHandler)
+	})
 }
 
 // go test -run Test_App_GETOnly