# Description

- Micro performance improvements for the Fiber Router.
- Refactored `requestHandler` into two methods. One for CustomCtx and one for DefaultCtx.

## Improvements to RemoveEscapeChar()

### After:

```bash
ubuntu@ubuntu:~/Desktop/git/fiber$ go test -benchmem -run=^$ -bench ^Benchmark_Utils_RemoveEscapeChar$ github.com/gofiber/fiber/v3 -count=4
goos: linux
goarch: amd64
pkg: github.com/gofiber/fiber/v3
cpu: AMD Ryzen 7 7800X3D 8-Core Processor           
Benchmark_Utils_RemoveEscapeChar-4   	48737022	        21.87 ns/op	      16 B/op	       1 allocs/op
Benchmark_Utils_RemoveEscapeChar-4   	51360970	        21.88 ns/op	      16 B/op	       1 allocs/op
Benchmark_Utils_RemoveEscapeChar-4   	53682570	        21.97 ns/op	      16 B/op	       1 allocs/op
Benchmark_Utils_RemoveEscapeChar-4   	45811604	        21.91 ns/op	      16 B/op	       1 allocs/op
PASS
ok  	github.com/gofiber/fiber/v3	4.494s
```

### Before:

```bash
ubuntu@ubuntu:~/Desktop/git/fiber$ go test -benchmem -run=^$ -bench ^Benchmark_Utils_RemoveEscapeChar$ github.com/gofiber/fiber/v3 -count=4
goos: linux
goarch: amd64
pkg: github.com/gofiber/fiber/v3
cpu: AMD Ryzen 7 7800X3D 8-Core Processor           
Benchmark_Utils_RemoveEscapeChar-4   	34108017	        33.72 ns/op	      16 B/op	       1 allocs/op
Benchmark_Utils_RemoveEscapeChar-4   	31172214	        33.48 ns/op	      16 B/op	       1 allocs/op
Benchmark_Utils_RemoveEscapeChar-4   	34733971	        33.61 ns/op	      16 B/op	       1 allocs/op
Benchmark_Utils_RemoveEscapeChar-4   	34788264	        33.87 ns/op	      16 B/op	       1 allocs/op
PASS
ok  	github.com/gofiber/fiber/v3	4.707s
```

## Router Improvements

Benchmark between v2, main, and this PR

![image](https://github.com/user-attachments/assets/efd8166d-1e95-4b4f-a693-ab27437b1bd1)

## Type of change

- [x] Performance improvement (non-breaking change which improves efficiency)