@gaby 2024-12-26
main branch vs this branch

![image](https://github.com/user-attachments/assets/12099229-7ca5-4f02-a43c-f532c96c8c9b)