@gaby 2024-12-28
@ReneWerner87 @efectn v2 vs main vs this branch

![image](https://github.com/user-attachments/assets/efd8166d-1e95-4b4f-a693-ab27437b1bd1)